[feat] 소켓 분리
