+++ Android/FunBox/app/src/main/java/com/rpg/funbox/presentation/MapSocket.kt
@@ -1,0 +1,7 @@
+package com.rpg.funbox.presentation
+
+import com.rpg.funbox.presentation.map.SocketApplication
+
+object MapSocket {
+    val mSocket = SocketApplication.get()
+}
