+++ Android/FunBox/app/src/main/java/com/rpg/funbox/presentation/login/splash/SplashActivity.kt
@@ -18,8 +18,8 @@ class SplashActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        collectLatestUiEvent()
+        startActivity(Intent(this@SplashActivity, MainActivity::class.java))
+        //collectLatestUiEvent()
     }
 
     private fun collectLatestUiEvent() {
